Wherever I made relevant changes in my controllers and views, I added code comments to describe what was happening, and what changes I made.
In my controllers I went more into depth, outlining what was actually happening with the code, as its more complex.

As I deleted some of the fields and the BatchFees table I had while doing the Database standard. I deleted it as it didn't align with my use case. Some of the queries I had made during the database standard involved the BatchFees table and the BatchNotes field, both which are deleted now. So I made new raw sql queries of same difficulty.